Move the corresponding code from StudioAnalyzeReport.cs to ReportServicecs Move the corresponding code from ProjectInformation.cs to ProjectService.cs Add validation for the existence of the reports directory when loading external project. Add try-catch blocks
