+++ Export Analysis Reports/Sdl.Community.ExportAnalysisReports/Interfaces/IProjectService.cs
@@ -3,12 +3,16 @@ using System.ComponentModel;
 using System.Xml;
 using Sdl.Community.ExportAnalysisReports.Model;
 using Sdl.ProjectAutomation.Core;
+using Sdl.ProjectAutomation.FileBased;
+using Sdl.TranslationStudioAutomation.IntegrationApi;
 
 namespace Sdl.Community.ExportAnalysisReports.Interfaces
 {
 	public interface IProjectService
 	{
 		string ProjectsXmlPath { get; set; }
+
+		ProjectsController ProjectController { get; set; }
 
 		List<string> AddFilePaths(List<string> studioProjectsPaths);
 
@@ -19,5 +23,7 @@ namespace Sdl.Community.ExportAnalysisReports.Interfaces
 		void RemoveSingleFileProjectLanguages(Dictionary<string, bool> languagesDictionary, BindingList<LanguageDetails> languages);
 
 		void SetProjectLanguages(List<ProjectDetails> projectDetails, bool isChecked, string languageName);
+
+		ProjectDetails GetExternalProjectDetails(string path, string reportFolderPath);
 	}
 }
