+++ Export Analysis Reports/Sdl.Community.ExportAnalysisReports/pluginpackage.manifest.xml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <PluginPackage xmlns="http://www.sdl.com/Plugins/PluginPackage/1.0">
   <PlugInName>SDL Trados Studio – Export Analysis Reports</PlugInName>
-  <Version>3.0.2.7</Version>
+  <Version>3.0.2.8</Version>
   <Description>Plugin which exports Studio analysis report to CSV</Description>
   <Author>SDL Community Developers</Author>
 	<Include>
